requirements.txt
apache-airflow>=2.7.0
pandas
playwright
sqlalchemy